refactor: merge awakened Devour into Doom Awakened as an intrinsic modifier
Devour was a separate hidden, not-learnable ability whose only job was
maximizing devoured neutral abilities' level. It doesn't need its own
KV entry or awaken-config replacement slot; folding its modifier logic
into Doom Awakened's intrinsic modifier removes a redundant ability
registration while keeping native Devour untouched.

Also rewrites the Doom Awakened tooltips to match native Doom's style:
duration is now a standalone line (matching native doom_bringer_doom),
and the Scepter/Devour notes are split into separate paragraphs instead
of one run-on sentence.
